tests/spark_streaming_tester/requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
python-dotenv==1.0.0